fastapi/requirements.txt
fastapi
uvicorn
asyncpg
pytz
Minio
requests
python-multipart
python-jose[cryptography]
pydantic
starlette